+++ R/compounds-metfrag.R
@@ -77,6 +77,8 @@ unifyMFNames <- function(mfr)
                  PharmacoInfo = "pharmacoInfo",
                  SafetyInfo = "safetyInfo",
                  ToxicityInfo = "toxicityInfo",
+                 DisorderDisease = "disorderDisease",
+                 Identification = "identification",
                  KnownUse = "knownUse",
                  FPSum = "annoTypeCount", # Nov2019 version (rename to Jan2020 version)
                  AnnoTypeCount = "annoTypeCount", # Jan2020 version